crm-api/app/lead/requirements.txt
requests==2.29.0
sqlalchemy==2.0.22
psycopg2-binary==2.9.9
pytz==2023.4
types-pytz==2023.4.0.20240130
python-dateutil==2.8.2